requirements.txt
pypdf==5.1.0
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
chromadb==0.5.5
sentence-transformers==3.0.1
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.5
rich==13.8.1
streamlit==1.37.1
transformers==4.44.2
torch==2.3.1
